requirements.txt
# motiongpt
numpy==1.23.1
tensorboard
pytorch_lightning
torchmetrics
omegaconf
shortuuid
chumpy
transformers==4.42.4
diffusers
einops
wandb
rich
matplotlib

# for visualization
smplx==0.1.28
trimesh
joblib==1.2.0
h5py
scikit-image
spacy
ftfy
more-itertools
natsort
pyrender
moviepy
librosa
triangle

# EMAGE
ConfigArgParse
fasttext
h5py
imageio
ipython
librosa
lmdb
loguru
matplotlib
moviepy
gradio
fasttext-wheel
opencv_contrib_python
opencv_python
pandas
peakutils
ptflops
python_igraph
pyvirtualdisplay
PyYAML
replicate
scikit_learn
scipy
soundfile
termcolor
textgrid
tqdm
trimesh
wandb
pyglet
smplx
tensorboard
#pyrender
bert_score
jiwer



turbot5